requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
chromadb>=0.4.22
openai>=1.7.0
pypdf>=3.17.4
python-docx>=1.1.0
tiktoken>=0.5.2
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
